+++ XCDYouTubeKit Demo/iOS Demo/DemoAsynchronousViewController.h
@@ -1,8 +1,10 @@
 //
-//  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
+//  Copyright (c) 2013-2015 Cédric Luthi. All rights reserved.
 //
 
 @interface DemoAsynchronousViewController : UIViewController
+
+@property (nonatomic, weak) IBOutlet UITextField *apiKeyTextField;
 
 - (IBAction) play:(id)sender;