+++ XCDYouTubeKit/XCDYouTubeVideoPlayerViewController.h
@@ -1,5 +1,5 @@
 //
-//  Copyright (c) 2013-2014 Cédric Luthi. All rights reserved.
+//  Copyright (c) 2013-2015 Cédric Luthi. All rights reserved.
 //
 
 #import <MediaPlayer/MediaPlayer.h>
@@ -31,7 +31,7 @@ MP_EXTERN NSString *const XCDYouTubeVideoUserInfoKey;
  *
  *  Use UIViewController’s `presentMoviePlayerViewControllerAnimated:` method to play a YouTube video fullscreen.
  *
- *  Use the `presentInView:` method to play a YouTube video inline.
+ *  Use the `<presentInView:>` method to play a YouTube video inline.
  */
 @interface XCDYouTubeVideoPlayerViewController : MPMoviePlayerViewController
 
@@ -44,11 +44,11 @@ MP_EXTERN NSString *const XCDYouTubeVideoUserInfoKey;
 /**
  *  Initializes a YouTube video player view controller
  *
- *  @param videoIdentifier A 11 characters YouTube video identifier. If the video identifier is invalid (including nil) the `MPMoviePlayerPlaybackDidFinishNotification` will be posted with a `MPMovieFinishReasonPlaybackError` reason.
+ *  @param videoIdentifier A 11 characters YouTube video identifier. If the video identifier is invalid the `MPMoviePlayerPlaybackDidFinishNotification` will be posted with a `MPMovieFinishReasonPlaybackError` reason.
  *
  *  @return An initialized YouTube video player view controller with the specified video identifier.
  *
- *  @discussion You pass a nil videoIdentifier and set the `videoIdentifier` property later.
+ *  @discussion You can pass a nil *videoIdentifier* (or use the standard `init` method instead) and set the `<videoIdentifier>` property later.
  */
 - (instancetype) initWithVideoIdentifier:(NSString *)videoIdentifier __attribute__((objc_designated_initializer));
 
@@ -74,9 +74,9 @@ MP_EXTERN NSString *const XCDYouTubeVideoUserInfoKey;
  *
  *  Defaults to @[ XCDYouTubeVideoQualityHTTPLiveStreaming, @(XCDYouTubeVideoQualityHD720), @(XCDYouTubeVideoQualityMedium360), @(XCDYouTubeVideoQualitySmall240) ]
  *
- *  You should set this property right after calling the `initWithVideoIdentifier:` method. Setting this property to nil restores its default values.
+ *  You should set this property right after calling the `<initWithVideoIdentifier:>` method. Setting this property to nil restores its default values.
  *
- *  @see `XCDYouTubeVideoQuality`
+ *  @see XCDYouTubeVideoQuality
  */
 @property (nonatomic, copy) NSArray *preferredVideoQualities;